cade-hooking/requirements.txt
frida-tools
psutil
